backend/flaskProject/requirements.txt
flask
flask-cors
flasgger
mysql-connector-python
pyjwt